requirements.txt
python-dotenv
discord.py
requests
nba_api
pandas
pillow
cairosvg